backend/requirements.txt
huggingface_hub==0.25.2 #required for backend to work with Hugging Face
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
fastapi==0.116.1
idna==3.10
jwt==1.4.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
sniffio==1.3.1
starlette==0.47.1
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn[standard]==0.24.0
numpy
python-dotenv